+++ src/Screens/Profile/index.ts
@@ -1,0 +1,3 @@
+import ProfileContainer from './ProfileContainer';
+
+export default ProfileContainer;
